Fixed a bunch of stuff
Also broke satellites spawning in.. I think I may have fixed it but the mac has slowed to a stop and I'm starving. Test on next load up. I changed the Awake to Start in the SatScript. it was all i could see that was different. all sats are loading into the same place which points to the TLE being wrong.